Load the actionlog params from model instead of the helper
