+++ libraries/namespacemap.php
@@ -8,7 +8,6 @@
 
 defined('_JEXEC') or die;
 
-use Joomla\CMS\Filesystem\File;
 use Joomla\CMS\Filesystem\Folder;
 use Joomla\CMS\Log\Log;
 
@@ -142,7 +141,7 @@ class JNamespacePsr4Map
 		 */
 		$error_reporting = error_reporting(0);
 
-		if (!File::write($this->file, implode("\n", $content)))
+		if (!file_put_contents($this->file, implode("\n", $content)))
 		{
 			Log::add('Could not save ' . $this->file, Log::WARNING);
