+++ tests/Unit/Libraries/Cms/MVC/Model/DatabaseModelTest.php
@@ -14,8 +14,6 @@ use Joomla\CMS\MVC\Factory\MVCFactoryInterface;
 use Joomla\CMS\MVC\Model\BaseDatabaseModel;
 use Joomla\CMS\Table\Table;
 use Joomla\Database\DatabaseInterface;
-use Joomla\Database\DatabaseQuery;
-use Joomla\Database\QueryInterface;
 use Joomla\Tests\Unit\UnitTestCase;
 
 /**
@@ -211,26 +209,5 @@ class DatabaseModelTest extends UnitTestCase
 		};
 
 		$this->assertEquals(5, $model->_getListCount('query'));
-	}
-
-	/**
-	 * Returns a database query instance.
-	 *
-	 * @param   DatabaseInterface  $db  The database
-	 *
-	 * @return  QueryInterface
-	 *
-	 * @since   __DEPLOY_VERSION__
-	 */
-	private function getQueryStub(DatabaseInterface $db): QueryInterface
-	{
-		return new class($db) extends DatabaseQuery
-		{
-			public function groupConcat($expression, $separator = ',')
-			{}
-
-			public function processLimit($query, $limit, $offset = 0)
-			{}
-		};
 	}
 }
