+++ tests/Unit/bootstrap.php
@@ -125,5 +125,10 @@ spl_autoload_register([new \Joomla\CMS\Autoload\ClassLoader($loader), 'loadClass
 // Register the class aliases for Framework classes that have replaced their Platform equivalents
 require_once JPATH_LIBRARIES . '/classmap.php';
 
+// Load extension classes
+require_once JPATH_LIBRARIES . '/namespacemap.php';
+$extensionPsr4Loader = new \JNamespacePsr4Map;
+$extensionPsr4Loader->load();
+
 // Define the Joomla version if not already defined.
 defined('JVERSION') or define('JVERSION', (new \Joomla\CMS\Version)->getShortVersion());
